% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{clnTxt}
\alias{clnTxt}
\title{Clean text fields}
\usage{
clnTxt(x)
}
\arguments{
\item{x}{would be the field that wanted to be formatted.}
}
\description{
Clean text field so there are less issues when posting the data to PI.  This function will only leaves letters and numbers and remove multiple and trailing spaces and characters that are now UTF-8.
}
\examples{
x$field <- clnTxt(x$field) # You can opt to store your clean values on a new field
}
\keyword{clean}
\keyword{field}
\keyword{text}

