% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{exportPI}
\alias{exportPI}
\title{Export PI data to CSV}
\usage{
exportPI(data, json, output)
}
\description{
This function will export the clean JSON vector into a CSV file.
}
\examples{
exportPI(data, json, file_output) #create CSV based on data and json
}
\keyword{clean}
\keyword{csv}
\keyword{json}

