% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{fmtJSON}
\alias{fmtJSON}
\title{Format JSON field}
\usage{
fmtJSON(x, pname)
}
\arguments{
\item{x}{would be the field that wanted to be formatted.}
}
\description{
This function will format JSON field by making it uniform and removing "word_count_message".  This is a result form PI if the word count is too low.  It will also add the person name from the data frame.
}
\examples{
clean.json <- fmtJSON(x, pname="Joe Doe") # Format JSON by adding person name
}
\keyword{clean}
\keyword{format}
\keyword{json}

