% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{getPI}
\alias{getPI}
\title{Get PI data results}
\usage{
getPI(url, x, dump, ssl = FALSE)
}
\description{
This function will post the data from R into PI as JSON format and store it back to R.
User will have the capability of encrypting with server (which it is not default).  This would have to be enable since the script is configured to be run with Node.js locally.
}
\examples{
new.x1 <- getPI(url="http://www.example.com", x=data$field, dump=NULL) # store results into new var x2
new.x2 <- getPI(url="https://www.example.com", x=data$field, dump=NULL, ssl=TRUE) # store results into new var x2, but information is encrypted with server
}
\keyword{download}
\keyword{pi}
\keyword{rcurl}

