% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{getPI2}
\alias{getPI2}
\title{Get PI data results 2}
\usage{
getPI2(url, x, usr, pwd, dump, ssl = FALSE, win = FALSE, capath)
}
\description{
This function will post the data from R into PI as JSON format and store it back to R by using CURL natively from the system.
CURL will need to be installed and configured properly (specially on Windows) to make sure that the queries run successfully.
User will have the capability of encrypting with server (which it is not default).  This would have to be enable since the script is configured to be run with Node.js locally.
}
\examples{
new.x1 <- getPI2(url="http://www.example.com", x=data$field, dump=NULL) # store results into new var x2
new.x2 <- getPI2(url="https://www.example.com", x=data$field, dump=NULL, ssl=TRUE) # store results into new var x2, but information is encrypted with server
new.x3 <- getPI2(url="https://www.example.com", x=data$field, dump=NULL, win=TRUE) # store results into new var x3 and using windows
new.x4 <- getPI2(url="https://www.example.com", x=data$field, dump=NULL, win=TRUE, capath="C:/pathtoca/cacert.pem") # store results into new var x4 and using windows
}
\keyword{curl}
\keyword{download}
\keyword{native}
\keyword{pi}

