% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{pkgHC}
\alias{pkgHC}
\title{Package(s) Health Check}
\usage{
pkgHC()
}
\description{
This function runs proper packages install if they don't exist on the library and then load those packages.
}
\examples{
pkgHC()
}
\keyword{healthchecks}
\keyword{install}
\keyword{package}

