% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibmwatsonpi_func.R
\name{startPI}
\alias{startPI}
\title{Start Personality Insights (PI)}
\usage{
startPI(x)
}
\description{
---NOTE: For Windows user only---. Start PI server and don't wait for anything to finish.  After using this, the user needs to remember to run stopPI() to stop server.
}
\examples{
startPI(x="C:/local_path_to_watson-developer-cloud/personality-insights-nodejs")
}
\keyword{node.js}
\keyword{server}
\keyword{start}

