% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibmwatsonpi_func.R
\name{stopPI}
\alias{stopPI}
\title{Stop Personality Insights (PI)}
\usage{
stopPI()
}
\description{
---NOTE: For Windows user only---. Kill PI server by filtering to image name and prompt output whether the kill was successful or not.  Note that this will only work on Win7 of above where the "taskkill" command is available.  Also, the user will get feedback when the server has been fully stop.
}
\examples{
stopPI()
}
\keyword{server}
\keyword{stop}
\keyword{taskkill}
\keyword{win7}

