% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{exportPI}
\alias{exportPI}
\title{Export PI data to CSV}
\usage{
exportPI(x)
}
\description{
This function will export the clean JSON vector into a CSV file.
}
\examples{
exportPI(x) # create CSV based on x
}
\keyword{clean}
\keyword{csv}
\keyword{json}

