% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwatsonpi_func.R
\name{stopPI}
\alias{stopPI}
\title{Stop Personality Insights (PI)}
\usage{
stopPI()
}
\description{
---NOTE: For Windows and Linux user only---. Kill PI server by filtering to image name and prompt output whether the kill was successful or not.  Note that this will only work on Win7 of above where the "taskkill" command is available and on Linux where "pkill" command is available.
}
\examples{
stopPI()
}
\keyword{linux}
\keyword{pkill}
\keyword{server}
\keyword{stop}
\keyword{taskkill}
\keyword{win7}

